Latent defect harvest: 498 argument defects fixed -- and a NEW class 303 sites wide
Stage 02's macro signatures told us WHERE reconstructions disagree. They cannot
say which side is right. Every claim below is a read of app_update.bin.

TWO IMAGE-DERIVED FACTS DID THE WORK
1. Gate branches CANNOT differ. Detecting gate pairs directly in the image --
   two bl sites with different targets whose backward CFG walks resolve to the
   SAME r0-defining instruction -- gives 618/618 with identical arity and
   register set. The shipped code sets arguments up ONCE, BEFORE the gate
   branch. So a branch with zero arguments has provably dropped them.
2. The format string is an exact arity oracle. Resolving the ldr r0,[pc] literal
   and counting conversions gives arity = 1 + fmt_words. 850 distinct format
   strings, ZERO with more than one arity; validated against register evidence
   at 2,296/2,299 (the 3 exceptions are tool artifacts -- functions receiving
   args already in r1/r2).

FOUR-WAY SPLIT (the deliverable, not a raw "fixed N")
  class                      flagged  confirmed  false alarm  unsettled
  log-route disagreement     318*        250          4          68
  zero-arg variadic calls    244         244          0          36
  zero-arg assert expansion    4           4          0           0
  assert_post_action proto   185           0        353           0
  * Stage 02's 143 was an UNDERCOUNT of the same class: its detector only
    matched the braced if{...}else{...} form; the brace-less and early-return
    spellings bring it to 318.

TWO HYPOTHESES REFUTED, INCLUDING ONE OF OURS
- assert_post_action: Stage 02 read `eors r0,r0` at 0x7e2ec and suspected 185
  call sites of inventing arguments. REFUTED. All 353 shipped call sites (253
  app + 100 net) load r0 = a file path and r1 = a line number. The
  reconstructions were right; the callee simply discards them. No change made.
- A prior baseline error, self-caught: the first per-object comparison used
  /private/tmp/g1-s1-app and showed 40 unexplained objects. That is the Stage 01
  TRANSFORM tree, not the live tree. Corrected to g1-i40d-app; the number is 1.

A NEW DEFECT CLASS, LARGER THAN THE ONE WE WENT LOOKING FOR: 303 INVENTED
ARGUMENTS. Ghidra emitted every live caller-saved register as an argument --
*alternate_log in display_reflash_handler.c is literally the gate load read as a
4th argument. A COMPILER CANNOT FLAG TOO MANY ARGUMENTS, which is why this
survived every prior pass. Identified with full evidence and DELIBERATELY NOT
CORRECTED: some are volatile reads, so deleting them is an R8 (observable MMIO
ordering) decision that needs its own pass.

GATES
Build exit 0, 0 errors; nm -u 0; 0 duplicate globals; verify_data 995/995;
check_thread_create_stack_args 10/10; pin gates 627/0/0.
Flash 955,048 -> 956,368 B (+1,320); .text +1,332; .data/.bss/RAM delta 0;
headroom 26,160 B. NET UNTOUCHED -- zero net inputs modified, so the 225,581 B
freeze holds by construction.
Per-object 2,766/2,837 byte-identical; 71 differ, 70 of which are exactly the
sources edited. The 71st is uname.c.obj, already documented as build-unstable.

HONEST CAVEATS
- The four framebuffers were NOT run: Renode belonged to a concurrent agent. No
  claim is made. .text moved +1,332 B so code addresses shifted; RAM moved 0, so
  iteration-38 probe addresses stay valid. g_log_level's boot value was NOT
  established, so it is NOT claimed that these branches are dead in the traced
  window.
- key_event_thread.c's 36 LOG() uses have BOTH halves empty, so no donor exists
  and source order != address order. Rather than invent format strings they are
  contained behind two asm-renamed placeholders with a comment pointing at the
  report; the object is byte-identical.
- 4 JSON files and one file under recon/refactor/stage_99/tree/ were corrupted by
  an over-broad escape fix and ALL REVERTED; recon/refactor/ is clean (verified
  independently: every modified tracked .json parses).
- The _sym mirrors (not in the build) retain 53 zero-arg sites each; stale
  partial copies where no donor exists.

Detail: recon/analysis/latent_defect_harvest.md

+++ recon/app/src/FUN_0002f4a8.c
@@ -1,8 +1,8 @@
 /* Reconstructed FUN_0002f4a8 @ 0x2f4a8  (parity: 300/300 trials, PROVEN) */
 #include <stdint.h>
-extern void DEBUG_PRINT(unsigned, ...);
+extern void DEBUG_PRINT(unsigned long, ...);
 extern int FUN_000167a8(void);
-extern void FUN_00019c70(unsigned, ...);
+extern void FUN_00019c70(unsigned long, ...);
 extern int FUN_0008638c(int);
 extern void FUN_00086c04(uintptr_t, int, int);
 extern void FUN_00086c78(uintptr_t, int, int);
