+++ recon/app/src/FUN_00059d24.c
@@ -4,7 +4,7 @@
 extern int FUN_00059b94(uint32_t,uint32_t);
 extern void FUN_0007350c(uint32_t,uint32_t,int,int);
 extern void FUN_0007e2ec(uint32_t,int);
-extern void FUN_0007e2fa(uint32_t,...);
+extern void FUN_0007e2fa(unsigned long, ...);
 extern int FUN_00080f92(void);
 extern uint64_t FUN_000826e0(int,uint32_t);
 extern int FUN_0008270c(uint32_t);
