+++ recon/app/src/FUN_0005f24c.c
@@ -1,6 +1,6 @@
 /* Reconstructed FUN_0005f24c @ 0x5f24c  (parity: 60/60 trials, PROVEN) */
 #include <stdint.h>
-extern void FUN_0007e2fa(unsigned int,...);
+extern void FUN_0007e2fa(unsigned long, ...);
 extern void FUN_0007e2ec(unsigned int,...);
 extern void FUN_00086518(int,int);
 typedef void (*release_fn_t)(int object);
