+++ recon/app/src/master_display_thread.c
@@ -4,8 +4,8 @@
  */
 #include <stdint.h>
 
-extern void DEBUG_PRINT(uintptr_t, ...);
-extern void debug_print(uintptr_t, ...); /* FUN_00019c70 @ 0x00019c70 */
+extern void DEBUG_PRINT(unsigned long, ...);
+extern void debug_print(unsigned long, ...); /* FUN_00019c70 @ 0x00019c70 */
 extern void *memset_bytes(void *, int, uint32_t); /* FUN_00086c78 */
 extern void k_msleep(uint32_t);                   /* FUN_0007cb8e */
 extern void wait_for_event(uint32_t, uint32_t);   /* FUN_0007cb8a */
