+++ recon/app/src/set_brightness_to_panel_reg.c
@@ -1,8 +1,8 @@
 /* Reconstructed set_brightness_to_panel_reg @ 0x46e3c  (parity: 300/300 trials, PROVEN) */
 #include <stdint.h>
-extern void DEBUG_PRINT(uint32_t, ...);
+extern void DEBUG_PRINT(unsigned long, ...);
 extern void *FUN_000167a8(void);
-extern void FUN_00019c70(uint32_t, ...);
+extern void FUN_00019c70(unsigned long, ...);
 extern void FUN_0007d642(int, int);
 extern void FUN_0007d67e(unsigned int);
 extern void FUN_0007d696(int);
