+++ recon/app/src_sym/gui_4bit_bitmap_override.c
@@ -1,8 +1,8 @@
 /* Reconstructed gui_4bit_bitmap_override @ 0x43d78  (parity: 300/300 trials, PROVEN) */
 #include <stdint.h>
-extern void DEBUG_PRINT(uint32_t, ...);
+extern void DEBUG_PRINT(unsigned long, ...);
 extern void *get_device_info(void);
-extern void debug_print(uint32_t, ...);
+extern void debug_print(unsigned long, ...);
 extern int atomic_get_3_0(void);
 extern int resource_manger_get(int, uint32_t, int*, int*, uint32_t*, int);
 extern void reflash_fb_data_to_lcd(uint32_t, uint32_t, unsigned int, unsigned int, int, int);
