+++ recon/app/src_sym/ui_raster_height_task.c
@@ -1,8 +1,8 @@
 /* Reconstructed ui_raster_height_task @ 0x430c0  (parity: 300/300 trials, PROVEN) */
 #include <stdint.h>
-extern void DEBUG_PRINT(uint32_t, ...);
+extern void DEBUG_PRINT(unsigned long, ...);
 extern void *get_device_info(void);
-extern void debug_print(uint32_t, ...);
+extern void debug_print(unsigned long, ...);
 extern void ui_DashBoard_task(uint32_t, uint32_t, int);
 extern void cal_panel_canvas_coord(int*, int*);
 extern void gui_screen_clear(void);
