+++ recon/verified/src/FUN_00032b24.c
@@ -1,6 +1,6 @@
 /* Reconstructed FUN_00032b24 @ 0x32b24  (parity: 300/300 trials, PROVEN) */
 #include <stdint.h>
-extern void DEBUG_PRINT(unsigned int, ...);
+extern void DEBUG_PRINT(unsigned long, ...);
 extern int FUN_0007d16a(void*,void*);
 typedef int (*fp5)(void*,int,void*,void*,uint32_t);
 uint32_t FUN_00032b24(int param_1, char* param_2, int param_3, uint32_t* param_4, uint8_t* param_5, uint32_t param_6){
