+++ recon/verified/src/FUN_000336a0.c
@@ -1,7 +1,7 @@
 /* Reconstructed FUN_000336a0 @ 0x336a0  (parity: 300/300 trials, PROVEN) */
 
 #include <stdint.h>
-extern void DEBUG_PRINT(int,...);
+extern void DEBUG_PRINT(unsigned long, ...);
 extern int FUN_00071c20(unsigned,void*,int,int,int);
 typedef int (*fp1)(int,unsigned,int);
 typedef int (*fp2)(int,unsigned,int,int);
